refactor: cli
